Refactoring

+++ src/main/kotlin/dev/schlaubi/musicbot/core/io/DurationSerializer.kt
@@ -11,7 +11,7 @@ import kotlin.time.ExperimentalTime
 
 @OptIn(ExperimentalTime::class)
 object DurationSerializer : KSerializer<Duration> {
-    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("Duration", PrimitiveKind.INT)
+    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("Duration", PrimitiveKind.LONG)
 
     override fun deserialize(decoder: Decoder): Duration = Duration.milliseconds(decoder.decodeLong())
